Increased clairity of instructions
When I was following these instructions, I was not aware that NodeJS and NPM were required as it is not specified in the instructions. The setup script did an excellent job of making me aware of this. I also fixed a minor clarity issue. This should cause a slight incremental increase in the setup efficiency. 

This change is rather minor, but it is the first of far larger contributions to come.